+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/KubernetesClientBuilder.java
@@ -23,6 +23,7 @@ import io.fabric8.kubernetes.client.utils.Serialization;
 import java.io.InputStream;
 import java.lang.reflect.InvocationTargetException;
 import java.util.concurrent.Executor;
+import java.util.function.Consumer;
 import java.util.function.Supplier;
 
 /**
@@ -44,6 +45,7 @@ public class KubernetesClientBuilder {
   private HttpClient.Factory factory;
   private Class<KubernetesClient> clazz;
   private ExecutorSupplier executorSupplier;
+  private Consumer<HttpClient.Builder> builderConsumer;
 
   public KubernetesClientBuilder() {
     // basically the same logic as in KubernetesResourceUtil for finding list types
@@ -60,6 +62,10 @@ public class KubernetesClientBuilder {
     }
   }
 
+  KubernetesClientBuilder(Class<KubernetesClient> clazz) {
+    this.clazz = clazz;
+  }
+
   public KubernetesClient build() {
     if (config == null) {
       config = new ConfigBuilder().build();
@@ -68,13 +74,21 @@ public class KubernetesClientBuilder {
       if (factory == null) {
         return clazz.getConstructor(Config.class).newInstance(config);
       }
-      HttpClient client = factory.createHttpClient(config);
+      HttpClient client = getHttpClient();
       return clazz.getConstructor(HttpClient.class, Config.class, ExecutorSupplier.class).newInstance(client, config,
           executorSupplier);
     } catch (InstantiationException | IllegalAccessException | IllegalArgumentException | InvocationTargetException
         | NoSuchMethodException | SecurityException e) {
       throw KubernetesClientException.launderThrowable(e);
     }
+  }
+
+  HttpClient getHttpClient() {
+    HttpClient.Builder builder = factory.newBuilder(config);
+    if (this.builderConsumer != null) {
+      this.builderConsumer.accept(builder);
+    }
+    return builder.build();
   }
 
   public KubernetesClientBuilder withConfig(Config config) {
@@ -102,7 +116,7 @@ public class KubernetesClientBuilder {
    * calls and writing to streams.
    * <p>
    * Only override if you need more control over the number of task threads used by the kubernetes client.
-   * 
+   *
    * @return this builder
    */
   public KubernetesClientBuilder withTaskExecutor(Executor executor) {
@@ -117,11 +131,22 @@ public class KubernetesClientBuilder {
    * There will be a call to {@link ExecutorSupplier#onClose(Executor)} when a client is closed.
    * <p>
    * Only override if you need more control over the number of task threads used by the kubernetes client.
-   * 
+   *
    * @return this builder
    */
   public KubernetesClientBuilder withTaskExecutorSupplier(ExecutorSupplier executorSupplier) {
     this.executorSupplier = executorSupplier;
+    return this;
+  }
+
+  /**
+   * Provide additional configuration for the {@link HttpClient} that is created for this {@link KubernetesClient}.
+   *
+   * @param consumer to modify the {@link HttpClient.Builder}
+   * @return this builder
+   */
+  public KubernetesClientBuilder withHttpClientBuilderConsumer(Consumer<HttpClient.Builder> consumer) {
+    this.builderConsumer = consumer;
     return this;
   }